requirement.txt
azure-ai-projects
azure-identity
openai
azure-ai-inference
python-dotenv
faiss-cpu
numpy
azure-ai-contentsafety
azure-ai-vision-imageanalysis
azure-ai-vision-face
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-customvision
pillow
matplotlib
azure-ai-textanalytics
azure-ai-translation-text
azure-cognitiveservices-speech